Test: Add st20p drop-when-late acceptance test via RxTxApp
Verify ST20P_TX_FLAG_DROP_WHEN_LATE end-to-end: start the user-pacing
clock a second in the past so every frame is late, then assert the TX
pipeline drops at least one frame-rate's worth of frames.

The test is framework-neutral in the style of test_fps: it is parametrized
by application and reads the drop count through a new app-agnostic
Application.count_tx_dropped_frames() helper, which parses the library's
TX pipeline stats line (emitted regardless of which framework drives MTL).
Only rxtxapp is listed today since the FFmpeg plugin has no drop-when-late
control; it can join the parametrization once it does.

RxTxApp already exposed user_pacing/drop_when_late in JSON; make its
user-pacing clock offset signed (int64) so a negative user_time_offset
starts the schedule in the past, and wire the three knobs through the
validation engine's st20p config builder.

+++ tests/tools/RxTxApp/src/args.c
@@ -949,7 +949,7 @@ int st_app_parse_args(struct st_app_context* ctx, struct mtl_init_params* p, int
         p->flags |= MTL_FLAG_NO_MULTICAST;
         break;
       case ST_ARG_TX_USER_CLOCK_OFFSET:
-        ctx->user_time.user_time_offset = atoi(optarg);
+        ctx->user_time.user_time_offset = atoll(optarg);
         break;
       case ST_ARG_AUTO_STOP:
         ctx->auto_stop = true;
